Word WILL NOW SERIALIZE!!!
made a backup of the dictionary.cs file named dictBACKUP.cs and marked
in properties to not compile. overhauled Dictionary.cs to be compatible
with .net XML serialization Engine.
updated WORDmaker to serialize Word class. however does not yet append
it is just for testing the produced schema of generic XML serialization.
deleted AssemblyInfo1 since all classes are public now.